Support for basic styles and borders
The BackBreeze library is intended to be higher level than Termite,
allowing for rich styling. Whilst many of the styles mirror those
provided by termite, more higher level styles are planned for the
future.

The first custom style is a border. The borders support full borders, or
partial, filling in the corners correctly when 2 or more borders are
set.