Add box rendering to allow a higher level API
Boxes can now be rendered. This is essentially a tree, which will render
the children and work their way back to the parent. This allows the
parent to know the width of its children which will be useful in working
out a layout.

Currently the children are all joined horizontally with no overflow,
this will be handled in a future commit.