Handle absolute positioning on boxes
Boxes can now be absolutely positioned relative to their parent. This
introduces a lot of flexibility and can be used to overlay certain
elements such as a panel title.

Whilst this could be achieved by using cursor escape codes, it makes
rendering things like tabs more difficult of there are cursor escape
codes in the middle of the output. Instead, a layering approach is used,
where the absolute layers are rendered based on their hierarchy. All
relative contents are currently squashed into a single layer, based on
the first time a relative item appears.

A layer is generated for the children, with the highest layer taking
priority, i.e. something absolutely positioned will render on top of the
existing contents.

A separate layer is generated for the parent, and also looked up at the
point of rendering. To ensure that styles persist, the escape code is
stored next to the character in the layer. These are grouped when
outputting so that if there are multiple characters with the same style,
there won't be an escape sequence per character.

The ucwidth library has also been included, to ensure correctness when
calculating the width for strings containing unicode.