feat(box): support display grid
Initial grid support is available, currently only supporting specifying
the number of columns. Grids are rendered based on the specified width
and height. To allow the screen width/height to be used, a helper
function has been added to the base module, either using the specified
terminal width or heigth, or calculating it if no terminal is specified.

To allow nested grids, the width/height have to be precomputed to
determine the required width/height of the nested grid, as boxes are
rendered from the innermost child outwards.