feat(style): allow setting border color
The border color can now be set. This is set on the style and then
passed through to the border at the point of rendering. Keeping it out
of the border struct during the population of the style struct means
that the state of the border struct doesn't need to be inspected, it can
simply be repopulated each time the border is changed.